+++ colang-lang/CHANGELOG.md
@@ -11,3 +11,19 @@ All notable changes to the "colang-lang" extension will be documented in this fi
 - Go to References
 - Find All References
 - Shortcut to comment code
+
+## [0.0.1]
+
+### Added
+- Rename symbol
+
+## [0.1.0]
+
+### Added
+- Syntax checker
+  - Unused symbol
+  - Duplicate symbol
+  - Undefined symbol
+
+### Fixed
+- Rename symbol